T: add way to test both variants of formatting with disabled and enabled feature in a single test
