--- src/test/resources/org/rust/ide/formatter/fixtures/align_params_in_calls_off_after.rs
@@ -1,19 +1,0 @@
-fn main() {
-    foooo(123,
-        456,
-        789);
-
-    write!(123,
-        456,
-        789);
-
-    let fooooooo = (123,
-        456,
-        789);
-
-    match 1 {
-        FooSome(1234, true,
-            5678, false,
-            91011, "foobar") => 10
-    }
-}
