--- src/test/resources/org/rust/ide/formatter/fixtures/align_params_off_after.rs
@@ -1,13 +1,0 @@
-enum Foooooooooo {
-    Soooome(i32,
-        i32,
-        i32)
-}
-
-fn fooooooooo(x: i32,
-    y: i32,
-    z: i32) {
-    let fooooooo: (i32,
-        i32,
-        i32) = ();
-}
