--- src/test/resources/org/rust/ide/formatter/fixtures/align_return_type.rs
@@ -1,28 +1,0 @@
-fn foo(x: i32, y: String)
-    -> String {
-    y + x.to_string()
-}
-
-fn foo(x: i32,
-       y: String)
-   -> String {
-    y + x.to_string()
-}
-
-fn foo(
-    x: i32,
-    y: String)
-        -> String {
-    y + x.to_string()
-}
-
-fn foo(
-    x: i32,
-    y: String
-)             -> String {
-    y + x.to_string()
-}
-
-fn foo<T>(t: T)
-    -> T
-          where T: Clone { }
