--- src/test/resources/org/rust/ide/formatter/fixtures/align_type_params_on_after.rs
@@ -1,10 +1,0 @@
-struct Foo<T: A + B + 'c,
-           K: X + Y + 'z> {}
-
-struct Foo<'x: 'y + 'z,
-           K: X + Y + 'z> {}
-
-struct Foo<T> where T: for<'a,
-                           'b> Fn(i32) -> () {
-    a: T
-}
