--- src/test/resources/org/rust/ide/formatter/fixtures/align_where_on_after.rs
@@ -1,5 +1,0 @@
-fn fooooooooo(a: i32,
-              b: i32,
-              c: i32)
-              -> i32
-              where T: 'static {}
