--- src/test/resources/org/rust/ide/formatter/fixtures/allow_one_line_match_after.rs
@@ -1,3 +1,0 @@
-fn main() {
-    let x = match func() { Ok(v) => v.unwrap_or(0), Err(_) => () };
-}
