--- src/test/resources/org/rust/ide/formatter/fixtures/allow_one_line_match_off_after.rs
@@ -1,6 +1,0 @@
-fn main() {
-    let x = match func() {
-        Ok(v) => v.unwrap_or(0),
-        Err(_) => ()
-    };
-}
