--- src/test/resources/org/rust/ide/formatter/fixtures/argument_alignment_after.rs
@@ -1,22 +1,0 @@
-fn foo(x: i32,
-       y: i32,
-       z: i32) {}
-
-pub fn new<S>(shape: S,
-              material_idx: usize)
-              -> Primitive
-    where S: Shape + 'static {}
-
-fn main() {
-    fooooo(1,
-           2,
-           3)
-}
-
-extern {
-    fn variadic(a: i32,
-                b: i32,
-                c: i32,
-                ...)
-                -> i32;
-}
