--- src/test/resources/org/rust/ide/formatter/fixtures/chain_call_alignment_off_after.rs
@@ -1,8 +1,0 @@
-fn main() {
-    let foo = moo.boo().goo()
-        .foo().bar()
-        .map(|x| x.foo().doo()
-            .moo().boo()
-            .bar().baz())
-        .baz().moo();
-}
